S20241

+++ microservices/TripComparator/Application/BusinessObjects/BusPosition.cs
@@ -4,7 +4,7 @@ namespace Application.BusinessObjects;
 
 public class BusPosition : IBusPositionUpdated
 {
-    public required int Seconds { get; init; }
+    public required double Seconds { get; init; }
 
     public required string Message { get; init; }
 }
